chore: refactor resize handler and use polyfill
